requirements.txt
openai>=1.40.0
opencv-python>=4.9.0.80
pillow>=10.3.0
numpy>=1.26.4
pyyaml>=6.0.1
tqdm>=4.66.4